Bug: Fix command deserialization
Had to cast the (void*) returned by malloc in a couple of places, so I made sure to cast it to the proper types. I also added a check for a null input string. I also fixed a bug in the deserialization function that essentially incorrectly processed bulk strings after a null bulk string was encountered.